Added msg.nim and two more tests.
msg.nim add a response Queue field to msg so the sender can inform the
receiver where to send back responses. To do this Msg and Queue needed
to be defined so its available to the other modules. In particular,
Queue is a RootObj and MsgQueue uses it as its base object so we don't
have to expose the innards of a MsgQueue.

Probably want to change the casts from QueuePtr to MsgQueuePtr to switch
type statements, so we can "gracefully" detect illegal casts.

On performance it turns out the benchmark isn't being useful it self
as the timing information is just got to much jitter as it is not. But
the number of messages that ml1Pm in t4 can process in 30 seconds and
then varying the number of threads was interesting, it looks like as
currently implemented the maximum msgs in 30 seconds was > 4,000,000
which I saw with 3 or 4 threads. Not very good since my machine has
12 CPU's.